hs3/requirements.txt
pyelftools>=0.29
pyftdi>=0.54.0
